kernels: int64 row offsets in the fused CE (rows x vocab > 2^31 crashed)
The CE triton kernel computed the row base as program_id(0) * vocab in
int32. qwen3.5's 248,320 vocab crosses 2^31 elements at row 8,650 — a
bs16 s1k round (16,384 rows) faulted with cudaErrorIllegalAddress
during M5.2 profiling; bs8 (8,192 rows) sat 5% under the line, which is
why every prior sweep was clean. Promote the row index to int64 before
the multiply (both passes share the base expression).

Regression test: 8,704 x 248,320 (just past the boundary, ~8.7 GB,
skips under 10 GB free VRAM), fused loss finite + dlogits rows past the
line checked against per-row fp32 softmax reference.

Audit of the sibling kernels: rmsnorm/swiglu/rope/adamw offset products
stay 30-60x below 2^31 at any shape this project lowers (adamw's worst
is the 1.02B-element head object); CE was the only kernel multiplying a
ROW INDEX by a ~250k width.

+++ src/dataflow/tasks/kernels/cross_entropy.py
@@ -46,7 +46,10 @@ if triton is not None:
         logits_ptr, targets_ptr, dlogits_ptr, nll_ptr,
         n_rows, vocab, BLOCK: tl.constexpr,
     ):
-        row = tl.program_id(0)
+        # int64 row index: row * vocab overflows int32 once rows x vocab
+        # crosses 2^31 elements (qwen3.5's 248,320 vocab hits it at 8,650
+        # rows — a bs16 s1k round; bs8 sat 5% under the line)
+        row = tl.program_id(0).to(tl.int64)
         cols = tl.arange(0, BLOCK)
         base = logits_ptr + row * vocab
